fix(Menu): revert new import structure
