--- src/components/menus/Menu/index.ts
@@ -1,4 +1,0 @@
-export { Item } from "./Item"
-export { Content } from "./Content"
-export { Trigger } from "./Trigger"
-export { Menu as Root } from "./Menu"
